Restructure DB design, Changes all the required entities and services

+++ src/main/java/in/lekhai/error/controller/role/exception/RoleForCategoryDoesNotExistException.java
@@ -4,7 +4,7 @@ import in.lekhai.core.enums.Roles;
 import in.lekhai.error.controller.LekhaiException;
 
 public class RoleForCategoryDoesNotExistException extends LekhaiException {
-    public RoleForCategoryDoesNotExistException(Roles role, Long category) {
+    public RoleForCategoryDoesNotExistException(Roles role, Integer category) {
         super(String.format("[%s] roles does not exist for category [%s]", role, category));
     }
 }
